+++ src/commands-helpers/workspace/create.helper.ts
@@ -1,13 +1,12 @@
 import { ListrTask } from 'listr2'
 import { fsWorkspaceManager, PMACMap } from '../../file-system'
 import { WorkspaceType, WorkspaceTypeValues } from '../../postman/api/types'
-import { PMACWorkspace } from '../../file-system/types'
+import { SharedWorkspacesTasksCtx } from '../shared/workspace-tasks/workspace-tasks-context.interface'
 
-export interface PmacWorkspaceCreateCtx {
+export interface PmacWorkspaceCreateCtx extends SharedWorkspacesTasksCtx.ICtxSelectedWorkspace {
     workspaceName: string
     workspaceType: WorkspaceType
     workspaceDescription?: string
-    workspaceJson?: PMACWorkspace
 }
 
 const setWorkspaceNameTask: ListrTask<PmacWorkspaceCreateCtx> = {
@@ -51,7 +50,7 @@ const setWorkspaceDescriptionTask: ListrTask<PmacWorkspaceCreateCtx> = {
 const buildWorkspaceJsonTask: ListrTask<PmacWorkspaceCreateCtx> = {
   title: 'Building a new workspace json',
   task: async ctx => {
-    ctx.workspaceJson = {
+    ctx.pmacWorkspace = {
       pmacID: PMACMap.generatePMACuuid(),
       pmID: '',
       pmUID: '',
@@ -69,22 +68,22 @@ const buildWorkspaceJsonTask: ListrTask<PmacWorkspaceCreateCtx> = {
 const createWorkspaceDirTask: ListrTask<PmacWorkspaceCreateCtx> =  {
   title: 'Creating a new directory for workspace',
   task: async ctx => {
-    if (!ctx.workspaceJson) {
+    if (!ctx.pmacWorkspace) {
       throw new Error('Cannot complete workspace directory creation when workspace json is not defined.')
     }
 
-    await fsWorkspaceManager.createPMACWorkspaceDir(ctx.workspaceJson)
+    await fsWorkspaceManager.createPMACWorkspaceDir(ctx.pmacWorkspace)
   },
 }
 
 const writeWorkspaceJsonToDirTask: ListrTask<PmacWorkspaceCreateCtx> = {
   title: 'Writing the new workspace json to workspace directory',
   task: async ctx => {
-    if (!ctx.workspaceJson) {
+    if (!ctx.pmacWorkspace) {
       throw new Error('Cannot complete workspace json file creation when workspace json is not defined.')
     }
 
-    await fsWorkspaceManager.writeWorkspaceDataJson(ctx.workspaceJson)
+    await fsWorkspaceManager.writeWorkspaceDataJson(ctx.pmacWorkspace)
   },
 }
 export const workspaceCreateCommandHelper = {