+++ src/commands-helpers/workspace/fetch.helper.ts
@@ -1,0 +1,3 @@
+import { SharedWorkspacesTasksCtx } from '../shared/workspace-tasks'
+
+export type PmacWorkspaceFetchCtx = SharedWorkspacesTasksCtx.TCtxWorkspacesCombined