+++ src/commands/workspace/push.ts
@@ -1,14 +1,12 @@
 import { Command } from '@oclif/core'
 
-import { PMACWorkspaceChooseAction, PMACWorkspaceGetAllAction, PMWorkspaceFetchAction, PMWorkspacePullToPMACAction } from '../../postman/actions'
+import { PMACWorkspaceChooseAction, PMACWorkspaceGetAllAction } from '../../postman/actions'
 import { postmanApiInstance } from '../../postman/api'
 import { fsWorkspaceManager, fsWorkspaceResourceManager } from '../../file-system'
-import { PostmanWorkspace } from '../../postman/api/types'
-import { PMACWorkspace } from '../../file-system/types'
 import inquirer from 'inquirer'
 import { PMWorkspaceCreateAction } from '../../postman/actions/pm-workspace-create.action'
 
-export default class WorkspaceFetchPulled extends Command {
+export default class WorkspacePush extends Command {
   static description = 'Fetches all pulled workspaces up-to-date.'
 
   static examples = [
@@ -17,11 +15,15 @@ export default class WorkspaceFetchPulled extends Command {
   ]
 
   async run(): Promise<void> {
-    await this.parse(WorkspaceFetchPulled)
+    await this.parse(WorkspacePush)
 
     const pmacWorkspaces = await new PMACWorkspaceGetAllAction(fsWorkspaceManager).run()
 
     const pmacWorkspace = await new PMACWorkspaceChooseAction(inquirer, pmacWorkspaces).run()
+
+    if (pmacWorkspace.pmacID) {
+      this.error(`Workspace '${pmacWorkspace.name}' already created in your Postman account.`)
+    }
 
     await new PMWorkspaceCreateAction(fsWorkspaceManager, fsWorkspaceResourceManager, postmanApiInstance, pmacWorkspace).run()